requirements.txt
attr==0.3.1
attrs==20.2.0
future==0.18.2
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
